Dev 20250325 multi keyboardlayout support (#78)
* Fix merge incorrect bug

* Dev 20250325 multi keyboardlayout support (#72)

Fix merge incorrect

* Fix merge incorrect

* Fix merge incorrect

* refactor(build-script): remove lupdate and lrelease validation check

The validation check for lupdate and lrelease binaries was removed as it is no longer necessary for the build process. This simplifies the script and reduces unnecessary complexity.

* style(build-script): fix spacing in conditional statement

Correct the spacing in the conditional statement to ensure proper syntax and improve readability

* Comments some log by writing file, change the language menu position

* refactor(settingdialog): remove redundant state and debug logging

Simplify the `changePage` method by removing the unnecessary `isChanging` flag and debug logging. This improves code readability and reduces complexity without altering the functionality.

* Add unicode mapping support and update keyboard layout configuration, add the document for the keyboard adding step